+++ website/components/CTAButtons.tsx
@@ -1,33 +1,33 @@
 "use client";
-import { FiCalendar, FiGithub } from "react-icons/fi";
+import { FiCalendar, FiTerminal } from "react-icons/fi";
 import styled from "styled-components";
 
 import { ButtonPrimary, ButtonTertiary } from "@/components/Button";
 import { mobile } from "@/constants/breakpoints";
 import { calendlyDemoUrl } from "@/constants/urls";
+import { scrollToId } from "@/utils";
 
 const CTAButtons = () => {
   return (
     <S.CTAButtons>
       <ButtonPrimary
-        analyticsId="button_hero_get_demo"
+        analyticsId="button_hero_get_started"
+        onClick={() => scrollToId("get-started")}
+        rel="noopener noreferrer"
+        target="_blank"
+        iconLeft={<FiTerminal size={18} />}
+        size="lg"
+      >
+        Get Started
+      </ButtonPrimary>
+      <ButtonTertiary
+        analyticsId="button_hero_schedule_demo"
         href={calendlyDemoUrl}
         rel="noopener noreferrer"
         target="_blank"
         iconLeft={<FiCalendar size={18} />}
-        size="lg"
       >
-        Get a Demo
-      </ButtonPrimary>
-      <ButtonTertiary
-        analyticsId="button_hero_github"
-        href="https://github.com/kurtosis-tech/kardinal"
-        rel="noopener noreferrer"
-        target="_blank"
-        iconLeft={<FiGithub size={18} />}
-        size="lg"
-      >
-        View on GitHub
+        Schedule a Demo
       </ButtonTertiary>
     </S.CTAButtons>
   );
@@ -39,7 +39,7 @@ namespace S {
     display: flex;
     flex-direction: row;
     align-items: center;
-    gap: 16px;
+    gap: 24px;
     @media ${mobile} {
       flex-direction: column;
     }